Struct cranelift_interpreter::interpreter::InterpreterState

pub struct InterpreterState<'a> {
    pub functions: FunctionStore<'a>,
    pub libcall_handler: LibCallHandler<DataValue>,
    pub frame_stack: Vec<Frame<'a>>,
    pub frame_offset: usize,
    pub stack: Vec<u8>,
    pub heaps: Vec<HeapBacking>,
    pub iflags: HashSet<IntCC>,
    pub fflags: HashSet<FloatCC>,
    pub pinned_reg: DataValue,
}

Maintains the Interpreter’s state, implementing the State trait.

Fields
functions: FunctionStore<'a>libcall_handler: LibCallHandler<DataValue>frame_stack: Vec<Frame<'a>>frame_offset: usize

Number of bytes from the bottom of the stack where the current frame’s stack space is

stack: Vec<u8>heaps: Vec<HeapBacking>iflags: HashSet<IntCC>fflags: HashSet<FloatCC>pinned_reg: DataValue
Implementations

impl<'a> InterpreterState<'a>

pub fn with_function_store(self, functions: FunctionStore<'a>) -> Self

pub fn with_libcall_handler(self, handler: LibCallHandler<DataValue>) -> Self

Registers a libcall handler

pub fn register_heap(&mut self, init: HeapInit) -> HeapId

Registers a static heap and returns a reference to it

This heap reference can be used to generate a heap pointer, which can be used inside the interpreter to load / store values into the heap.

let mut state = InterpreterState::default();
let heap0 = state.register_heap(HeapInit::Zeroed(1024));

let backing = Vec::from([10u8; 24]);
let heap1 = state.register_heap(HeapInit::FromBacking(backing));

pub fn get_heap_address(
&self,
ty: Type,
heap_id: HeapId,
offset: u64
) -> Result<DataValue, MemoryError>

Returns a heap address that can be used inside the interpreter

let mut state = InterpreterState::default();
let heap_id = state.register_heap(HeapInit::Zeroed(1024));
let heap_base = state.get_heap_address(I64, heap_id, 0);
let heap_bound = state.get_heap_address(I64, heap_id, 1024);