Module cranelift_interpreter::step

The step function interprets a single Cranelift instruction given its State and InstructionContext; the interpretation is generic over Values.

Enums

ControlFlow

Enumerate the ways in which the control flow can change based on a single step in a Cranelift interpreter.

CraneliftTrap

StepError

Functions

step

Interpret a single Cranelift instruction. Note that program traps and interpreter errors are distinct: a program trap results in Ok(Flow::Trap(...)) whereas an interpretation error (e.g. the types of two values are incompatible) results in Err(...).