Enum cranelift_interpreter::step::ControlFlow

pub enum ControlFlow<'a, V> {
    Assign(SmallVec<[V; 1]>),
    Continue,
    ContinueAt(Block, SmallVec<[V; 1]>),
    Call(&'a Function, SmallVec<[V; 1]>),
    Return(SmallVec<[V; 1]>),
    Trap(CraneliftTrap),
}

Enumerate the ways in which the control flow can change based on a single step in a Cranelift interpreter.

Variants

Assign(SmallVec<[V; 1]>)

Return one or more values from an instruction to be assigned to a left-hand side, e.g.: in v0 = iadd v1, v2, the sum of v1 and v2 is assigned to v0.

Tuple Fields of Assign

0: SmallVec<[V; 1]>

Continue

Continue to the next available instruction, e.g.: in nop, we expect to resume execution at the instruction after it.

ContinueAt(Block, SmallVec<[V; 1]>)

Jump to another block with the given parameters, e.g.: in brz v0, block42, [v1, v2], if the condition is true, we continue execution at the first instruction of block42 with the values in v1 and v2 filling in the block parameters.

Tuple Fields of ContinueAt

0: Block1: SmallVec<[V; 1]>

Call(&'a Function, SmallVec<[V; 1]>)

Indicates a call the given [Function] with the supplied arguments.

Tuple Fields of Call

0: &'a Function1: SmallVec<[V; 1]>

Return(SmallVec<[V; 1]>)

Return from the current function with the given parameters, e.g.: return [v1, v2].

Tuple Fields of Return

0: SmallVec<[V; 1]>

Trap(CraneliftTrap)

Stop with a program-generated trap; note that these are distinct from errors that may occur during interpretation.

Tuple Fields of Trap

0: CraneliftTrap

Implementations

impl<'a, V> ControlFlow<'a, V>

pub fn unwrap_return(self) -> Vec<V>

For convenience, we can unwrap the ControlFlow state assuming that it is a ControlFlow::Return, panicking otherwise.

pub fn unwrap_trap(self) -> CraneliftTrap

For convenience, we can unwrap the ControlFlow state assuming that it is a ControlFlow::Trap, panicking otherwise.