Crate cranelift_interpreter

Cranelift IR interpreter.

This module is a project for interpreting Cranelift IR.

Modules

address

Virtual Addressing Scheme for the Interpreter

environment

Implements the function environment (e.g. a name-to-function mapping) for interpretation.

frame

Implements a call frame (activation record) for the Cranelift interpreter.

instruction

The InstructionContext trait describes a Cranelift instruction; a default implementation is provided with DfgInstructionContext

interpreter

Cranelift IR interpreter.

state

Cranelift instructions modify the state of the machine; the State trait describes these ways this can happen.

step

The step function interprets a single Cranelift instruction given its State and InstructionContext.

value

The DataValueExt trait is an extension trait for DataValue. It provides a lot of functions used by the rest of the interpreter.