Crate cranelift_interpreter

Cranelift IR interpreter.

This module is a project for interpreting Cranelift IR.

Modules

address

Virtual Addressing Scheme for the Interpreter

environment

Implements the function environment (e.g. a name-to-function mapping) for interpretation.

frame

Implements a call frame (activation record) for the Cranelift interpreter.

instruction

The InstructionContext trait describes a Cranelift instruction; a default implementation is provided with DfgInstructionContext

interpreter

Cranelift IR interpreter.

state

Cranelift instructions modify the state of the machine; the State trait describes these ways this can happen.

step

The step function interprets a single Cranelift instruction given its State and InstructionContext; the interpretation is generic over Values.

value

The Value trait describes what operations can be performed on interpreter values. The interpreter usually executes using DataValues so an implementation is provided here. The fact that Value is a trait, however, allows interpretation of Cranelift IR on other kinds of values.