Module cranelift_interpreter::address

Virtual Addressing Scheme for the Interpreter

The interpreter uses virtual memory addresses for its memory operations. These addresses are obtained by the various _addr instructions (e.g. stack_addr) and can be either 32 or 64 bits.

Addresses are composed of 3 fields: “region”, “entry” and offset.

“region” refers to the type of memory that this address points to. “entry” refers to which instance of this memory the address points to (e.g table1 would be “entry” 1 of a Table region address). The last field is the “offset”, which refers to the offset within the entry.

The address has the “region” field as the 2 most significant bits. The following bits are the “entry” field, the amount of “entry” bits depends on the size of the address and the “region” of the address. The remaining bits belong to the “offset” field

An example address could be a 32 bit address, in the heap region, which has 2 “entry” bits this address would have 32 - 2 - 2 = 28 offset bits.

The only exception to this is the “stack” region, where, because we only have a single “stack” we have 0 “entry” bits, and thus is all offset.

address size	address kind	region value (2 bits)	entry bits (#)	offset bits (#)

32	Stack	0b00	0	30

32	Heap	0b01	2	28

32	Table	0b10	5	25

32	GlobalValue	0b11	6	24

64	Stack	0b00	0	62

64	Heap	0b01	6	56

64	Table	0b10	10	52

64	GlobalValue	0b11	12	50

Structs

Address

Enums

AddressRegion

Virtual Address region

AddressSize