Module cranelift_interpreter::instruction

The InstructionContext trait describes a Cranelift instruction; a default implementation is provided with DfgInstructionContext

Structs

DfgInstructionContext

Since InstructionContext is likely used within a Cranelift context in which a DataFlowGraph is available, a default implementation is provided–DfgInstructionContext.

Traits

InstructionContext

Exposes the necessary information for understanding a single Cranelift instruction. It would be nice if InstructionData contained everything necessary for interpreting the instruction, but Cranelift’s current design requires looking at other structures. A default implementation using a reference to a DataFlowGraph is provided in DfgInstructionContext.