Trait cranelift_interpreter::instruction::InstructionContext

pub trait InstructionContext {
    fn data(&self) -> InstructionData;

    fn args(&self) -> &[Value];

    fn type_of(&self, v: Value) -> Option<Type>;

    fn controlling_type(&self) -> Option<Type>;
}

Exposes the necessary information for understanding a single Cranelift instruction. It would be nice if InstructionData contained everything necessary for interpreting the instruction, but Cranelift's current design requires looking at other structures. A default implementation using a reference to a DataFlowGraph is provided in DfgInstructionContext.

Required methods

fn data(&self) -> InstructionData

fn args(&self) -> &[Value]

fn type_of(&self, v: Value) -> Option<Type>

fn controlling_type(&self) -> Option<Type>

Loading content...