Module cranelift_interpreter::interpreter

Cranelift IR interpreter.

This module partially contains the logic for interpreting Cranelift IR.

Structs

HeapId

Represents a registered heap with an interpreter.

Interpreter

The Cranelift interpreter; this contains some high-level functions to control the interpreter’s flow. The interpreter state is defined separately (see InterpreterState) as the execution semantics for each Cranelift instruction (see step).

InterpreterState

Maintains the Interpreter’s state, implementing the State trait.

Enums

FuelResult

The result of consuming fuel. Signals if the caller should stop or continue.

HeapInit

Options for initializing a heap memory region

InterpreterError

The ways interpretation can fail.

Type Definitions

HeapBacking

LibCallHandler

LibCallValues