Trait cranelift_interpreter::state::State

pub trait State<'a, V> {

    // Required methods
    fn get_function(&self, func_ref: FuncRef) -> Option<&'a Function>;
    fn get_current_function(&self) -> &'a Function;
    fn get_libcall_handler(&self) -> LibCallHandler<V>;
    fn push_frame(&mut self, function: &'a Function);
    fn pop_frame(&mut self);
    fn get_value(&self, name: Value) -> Option<V>;
    fn set_value(&mut self, name: Value, value: V) -> Option<V>;
    fn stack_address(
        &self,
        size: AddressSize,
        slot: StackSlot,
        offset: u64
    ) -> Result<Address, MemoryError>;
    fn checked_load(
        &self,
        address: Address,
        ty: Type,
        mem_flags: MemFlags
    ) -> Result<V, MemoryError>;
    fn checked_store(
        &mut self,
        address: Address,
        v: V,
        mem_flags: MemFlags
    ) -> Result<(), MemoryError>;
    fn function_address(
        &self,
        size: AddressSize,
        name: &ExternalName
    ) -> Result<Address, MemoryError>;
    fn get_function_from_address(
        &self,
        address: Address
    ) -> Option<InterpreterFunctionRef<'a>>;
    fn resolve_global_value(&self, gv: GlobalValue) -> Result<V, MemoryError>;
    fn get_pinned_reg(&self) -> V;
    fn set_pinned_reg(&mut self, v: V);

    // Provided method
    fn collect_values(&self, names: &[Value]) -> Result<SmallVec<[V; 1]>, Value> { ... }

}

This trait manages the state necessary to interpret a single Cranelift instruction–it describes all of the ways a Cranelift interpreter can interact with its virtual state. This makes it possible to use the Interpreter in a range of situations:

when interpretation requires understanding all of the ways state can change (e.g. loading and storing from the heap) we will use a full-fledged state, like InterpreterState.

when interpretation can ignore some state changes (e.g. abstract interpretation of arithmetic instructions–no heap knowledge required), we can partially implement this trait. See ImmutableRegisterState for an example of this: it only exposes the values referenced by the SSA references in the current frame and not much else.

Required Methods

fn get_function(&self, func_ref: FuncRef) -> Option<&'a Function>

Retrieve a reference to a Function.

fn get_current_function(&self) -> &'a Function

Retrieve a reference to the currently executing Function.

fn get_libcall_handler(&self) -> LibCallHandler<V>

Retrieve the handler callback for a LibCall

fn push_frame(&mut self, function: &'a Function)

Record that an interpreter has called into a new Function.

fn pop_frame(&mut self)

Record that an interpreter has returned from a called Function.

fn get_value(&self, name: Value) -> Option<V>

Retrieve a value V by its value reference from the virtual register file.

fn set_value(&mut self, name: Value, value: V) -> Option<V>

Assign a value V to its value reference in the virtual register file.

fn stack_address( &self, size: AddressSize, slot: StackSlot, offset: u64 ) -> Result<Address, MemoryError>

Computes the stack address for this stack slot, including an offset.

fn checked_load( &self, address: Address, ty: Type, mem_flags: MemFlags ) -> Result<V, MemoryError>

Retrieve a value V from memory at the given address, checking if it belongs either to the stack or to one of the heaps; the number of bytes loaded corresponds to the specified Type.

fn checked_store( &mut self, address: Address, v: V, mem_flags: MemFlags ) -> Result<(), MemoryError>

Store a value V into memory at the given address, checking if it belongs either to the stack or to one of the heaps; the number of bytes stored corresponds to the specified Type.

fn function_address( &self, size: AddressSize, name: &ExternalName ) -> Result<Address, MemoryError>

Compute the address of a function given its name.

fn get_function_from_address( &self, address: Address ) -> Option<InterpreterFunctionRef<'a>>

Retrieve a reference to a Function given its address.

fn resolve_global_value(&self, gv: GlobalValue) -> Result<V, MemoryError>

Given a global value, compute the final value for that global value, applying all operations in intermediate global values.

fn get_pinned_reg(&self) -> V

Retrieves the current pinned reg value

fn set_pinned_reg(&mut self, v: V)

Sets a value for the pinned reg

Provided Methods

fn collect_values(&self, names: &[Value]) -> Result<SmallVec<[V; 1]>, Value>

Collect a list of values V by their value references; this is a convenience method for get_value. If no value is found for a value reference, return an Err containing the offending reference.