Trait cranelift_interpreter::state::State

pub trait State<'a, V> {

    fn get_function(&self, func_ref: FuncRef) -> Option<&'a Function>;

    fn push_frame(&mut self, function: &'a Function);

    fn pop_frame(&mut self);

    fn get_value(&self, name: Value) -> Option<V>;

    fn set_value(&mut self, name: Value, value: V) -> Option<V>;

    fn has_iflag(&self, flag: IntCC) -> bool;

    fn set_iflag(&mut self, flag: IntCC);

    fn has_fflag(&self, flag: FloatCC) -> bool;

    fn set_fflag(&mut self, flag: FloatCC);

    fn clear_flags(&mut self);

    fn load_heap(&self, offset: usize, ty: Type) -> Result<V, MemoryError>;

    fn store_heap(&mut self, offset: usize, v: V) -> Result<(), MemoryError>;

    fn load_stack(&self, offset: usize, ty: Type) -> Result<V, MemoryError>;

    fn store_stack(&mut self, offset: usize, v: V) -> Result<(), MemoryError>;

    fn collect_values(&self, names: &[Value]) -> Result<SmallVec<[V; 1]>, Value> { ... }

}

This trait manages the state necessary to interpret a single Cranelift instruction–it describes all of the ways a Cranelift interpreter can interact with its virtual state. This makes it possible to use the Interpreter in a range of situations:

when interpretation requires understanding all of the ways state can change (e.g. loading and storing from the heap) we will use a full-fledged state, like InterpreterState.

when interpretation can ignore some state changes (e.g. abstract interpretation of arithmetic instructions–no heap knowledge required), we can partially implement this trait. See ImmutableRegisterState for an example of this: it only exposes the values referenced by the SSA references in the current frame and not much else.

Required methods

fn get_function(&self, func_ref: FuncRef) -> Option<&'a Function>

Retrieve a reference to a Function.

fn push_frame(&mut self, function: &'a Function)

Record that an interpreter has called into a new Function.

fn pop_frame(&mut self)

Record that an interpreter has returned from a called Function.

fn get_value(&self, name: Value) -> Option<V>

Retrieve a value V by its value reference from the virtual register file.

fn set_value(&mut self, name: Value, value: V) -> Option<V>

Assign a value V to its value reference in the virtual register file.

fn has_iflag(&self, flag: IntCC) -> bool

Check if an IntCC flag has been set.

fn set_iflag(&mut self, flag: IntCC)

Set an IntCC flag.

fn has_fflag(&self, flag: FloatCC) -> bool

Check if a FloatCC flag has been set.

fn set_fflag(&mut self, flag: FloatCC)

Set a FloatCC flag.

fn clear_flags(&mut self)

Clear all IntCC and FloatCC flags.

fn load_heap(&self, offset: usize, ty: Type) -> Result<V, MemoryError>

Retrieve a value V from the heap at the given offset; the number of bytes loaded corresponds to the specified Type.

fn store_heap(&mut self, offset: usize, v: V) -> Result<(), MemoryError>

Store a value V into the heap at the given offset. The Type of V will determine the number of bytes stored.

fn load_stack(&self, offset: usize, ty: Type) -> Result<V, MemoryError>

Retrieve a value V from the stack at the given offset; the number of bytes loaded corresponds to the specified Type.

fn store_stack(&mut self, offset: usize, v: V) -> Result<(), MemoryError>

Store a value V on the stack at the given offset. The Type of V will determine the number of bytes stored.

Provided methods

fn collect_values(&self, names: &[Value]) -> Result<SmallVec<[V; 1]>, Value>

Collect a list of values V by their value references; this is a convenience method for get_value. If no value is found for a value reference, return an Err containing the offending reference.